Add files via upload
This R script will take excel files from a plate reader and output a graph plotting polarization shifts against the base 10 logarithm of the concentration range tested. This script is designed for FP competition assays but should also work for binding assays.

The libraries needed for this script are tidyverse and readxl.

This script plots triplicate data where all 3 replicates are scanned 3 times and therefore requires 3 excel files. Specifically, the script is set up to extract data from a 96 well plate where B1 has the probe only control and columns, 6, 7, and 8 in rows A to H have the dilutions of receptor or ligand (each replicate is grouped by column).

Specify the cold ligand highest concentration and the dilution factor for the experiment.